pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "Jarvis - 基于契约驱动的智能任务编排系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "tools*", "skills*", "memory*", "apps*"]
